move production dataclasses to subfolder
